Módulo calificaciones

+++ database/migrations/2026_03_05_225506_create_calificacions_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('grupo_id')->constrained();
             $table->foreignId('user_id')->constrained();
-            $table->float('calificacion');
+            $table->decimal('calificacion', 8, 2);
             $table->timestamps();
         });
     }
